+++ src/commands/cache.rs
@@ -8,8 +8,8 @@ use clap::{Parser, Subcommand};
 use serde::Serialize;
 use tokio::task::JoinSet;
 
-mod list;
-mod remove;
+pub mod list;
+pub mod remove;
 
 use crate::{abandon, commands::CommandRunner, matching::PatternMatcher, wrappers::sizes::Size};